Stage 8 - Functioning Aliens & Barriers
I am happy with the progress so far. The aliens now hold formation and move as one reacting to if they hit a wall. In the next update I will be working on moving the aliens down with each wall touch. The barriers now also work and the user can wear them down if accidentally shot.