Stage 10 - Complete Project
The game is complete! The game restarts when the user is shot or the aliens reach the bottom. The game runs smoothly and I am very happy. Possibly in the future I will come back to this app and make changes but we will have to see, for right now I am happy with what I have created, not bad for a few days!